Refactor code to reuse and fix error of wrong cmdArgs
