+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/task/RunNativeImageTestTask.java
@@ -20,6 +20,7 @@ package io.ballerina.cli.task;
 
 import io.ballerina.cli.utils.BuildTime;
 import io.ballerina.cli.utils.GraalVMCompatibilityUtils;
+import io.ballerina.cli.utils.NativeUtils;
 import io.ballerina.cli.utils.TestUtils;
 import io.ballerina.projects.JBallerinaBackend;
 import io.ballerina.projects.JarResolver;
@@ -34,70 +35,39 @@ import io.ballerina.projects.ProjectException;
 import io.ballerina.projects.ProjectKind;
 import io.ballerina.projects.internal.model.Target;
 import io.ballerina.projects.util.ProjectConstants;
-import org.apache.commons.compress.utils.IOUtils;
-import org.ballerinalang.test.runtime.entity.MockFunctionReplaceVisitor;
 import org.ballerinalang.test.runtime.entity.ModuleStatus;
 import org.ballerinalang.test.runtime.entity.TestReport;
 import org.ballerinalang.test.runtime.entity.TestSuite;
 import org.ballerinalang.test.runtime.util.TesterinaConstants;
-import org.ballerinalang.test.runtime.util.TesterinaUtils;
 import org.ballerinalang.testerina.core.TestProcessor;
-import org.objectweb.asm.ClassReader;
-import org.objectweb.asm.ClassVisitor;
-import org.objectweb.asm.ClassWriter;
-import org.objectweb.asm.Opcodes;
-import org.objectweb.asm.Type;
 import org.wso2.ballerinalang.util.Lists;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.PrintStream;
-import java.lang.reflect.Method;
-import java.net.URL;
-import java.net.URLClassLoader;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.security.AccessController;
-import java.security.PrivilegedAction;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Scanner;
-import java.util.StringJoiner;
-import java.util.jar.JarOutputStream;
-import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
 import static io.ballerina.cli.launcher.LauncherUtils.createLauncherException;
-import static io.ballerina.cli.utils.NativeUtils.createReflectConfig;
-import static io.ballerina.cli.utils.NativeUtils.getURLList;
 import static io.ballerina.cli.utils.TestUtils.generateTesterinaReports;
 import static io.ballerina.projects.util.ProjectConstants.BIN_DIR_NAME;
-import static io.ballerina.runtime.api.constants.RuntimeConstants.FILE_NAME_PERIOD_SEPARATOR;
-import static java.util.Objects.requireNonNull;
+import static io.ballerina.projects.util.ProjectConstants.TESTS_CACHE_DIR_NAME;
 import static org.ballerinalang.test.runtime.util.TesterinaConstants.CACHE_DIR;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.CLASS_EXTENSION;
 import static org.ballerinalang.test.runtime.util.TesterinaConstants.DOT;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.DOT_REPLACER;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.HYPHEN;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.JAR_EXTENSION;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.JAVA_17_DIR;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.MOCK_FN_DELIMITER;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.MOCK_FUNC_NAME_PREFIX;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.MOCK_LEGACY_DELIMITER;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.MODIFIED;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.PATH_SEPARATOR;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.TESTABLE;
+import static org.ballerinalang.test.runtime.util.TesterinaConstants.TESTERINA_TEST_SUITE;
 
 /**
  * Task for executing tests.
@@ -105,8 +75,6 @@ import static org.ballerinalang.test.runtime.util.TesterinaConstants.TESTABLE;
  * @since 2.3.0
  */
 public class RunNativeImageTestTask implements Task {
-
-    private static final String OS = System.getProperty("os.name").toLowerCase(Locale.getDefault());
     private static final String WIN_EXEC_EXT = "exe";
 
     private static class StreamGobbler extends Thread {
@@ -156,124 +124,6 @@ public class RunNativeImageTestTask implements Task {
         }
         this.listGroups = listGroups;
         this.isParallelExecution = isParallelExecution;
-    }
-
-
-    private static byte[] getModifiedClassBytes(String className, List<String> functionNames, TestSuite suite,
-                                                ClassLoader classLoader) {
-        Class<?> functionToMockClass;
-        try {
-            functionToMockClass = classLoader.loadClass(className);
-        } catch (Throwable e) {
-            throw createLauncherException("failed to load class: " + className);
-        }
-
-        byte[] classFile = new byte[0];
-        boolean readFromBytes = false;
-        for (Method method1 : functionToMockClass.getDeclaredMethods()) {
-            if (functionNames.contains(MOCK_FN_DELIMITER + method1.getName())) {
-                String desugaredMockFunctionName = MOCK_FUNC_NAME_PREFIX + method1.getName();
-                String testClassName = TesterinaUtils.getQualifiedClassName(suite.getOrgName(),
-                        suite.getTestPackageID(), suite.getVersion(),
-                        suite.getPackageID().replace(DOT, FILE_NAME_PERIOD_SEPARATOR));
-                Class<?> testClass;
-                try {
-                    testClass = classLoader.loadClass(testClassName);
-                } catch (Throwable e) {
-                    throw createLauncherException("failed to prepare " + testClassName + " for mocking reason:" +
-                            e.getMessage());
-                }
-                for (Method method2 : testClass.getDeclaredMethods()) {
-                    if (method2.getName().equals(desugaredMockFunctionName)) {
-                        if (!readFromBytes) {
-                            classFile = replaceMethodBody(method1, method2);
-                            readFromBytes = true;
-                        } else {
-                            classFile = replaceMethodBody(classFile, method1, method2);
-                        }
-                    }
-                }
-            } else if (functionNames.contains(MOCK_LEGACY_DELIMITER + method1.getName())) {
-                String key = className + MOCK_LEGACY_DELIMITER + method1.getName();
-                String mockFunctionName = suite.getMockFunctionNamesMap().get(key);
-                if (mockFunctionName != null) {
-                    String mockFunctionClassName = suite.getTestUtilityFunctions().get(mockFunctionName);
-                    Class<?> mockFunctionClass;
-                    try {
-                        mockFunctionClass = classLoader.loadClass(mockFunctionClassName);
-                    } catch (ClassNotFoundException e) {
-                        throw createLauncherException("failed to prepare " + mockFunctionClassName +
-                                " for mocking reason:" + e.getMessage());
-                    }
-                    for (Method method2 : mockFunctionClass.getDeclaredMethods()) {
-                        if (method2.getName().equals(mockFunctionName)) {
-                            if (!readFromBytes) {
-                                classFile = replaceMethodBody(method1, method2);
-                                readFromBytes = true;
-                            } else {
-                                classFile = replaceMethodBody(classFile, method1, method2);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        return classFile;
-    }
-
-    private static byte[] replaceMethodBody(Method method, Method mockMethod) {
-        Class<?> clazz = method.getDeclaringClass();
-        ClassReader cr;
-        try {
-            InputStream ins;
-            ins = clazz.getResourceAsStream(clazz.getSimpleName() + CLASS_EXTENSION);
-            cr = new ClassReader(requireNonNull(ins));
-        } catch (IOException e) {
-            throw createLauncherException("failed to get the class reader object for the class "
-                    + clazz.getSimpleName());
-        }
-        ClassWriter cw = new ClassWriter(cr, ClassWriter.COMPUTE_FRAMES | ClassWriter.COMPUTE_MAXS);
-        ClassVisitor cv = new MockFunctionReplaceVisitor(Opcodes.ASM7, cw, method.getName(),
-                Type.getMethodDescriptor(method), mockMethod);
-        cr.accept(cv, 0);
-        return cw.toByteArray();
-    }
-
-    private static byte[] replaceMethodBody(byte[] classFile, Method method, Method mockMethod) {
-        ClassReader cr = new ClassReader(classFile);
-        ClassWriter cw = new ClassWriter(cr, ClassWriter.COMPUTE_FRAMES | ClassWriter.COMPUTE_MAXS);
-        ClassVisitor cv = new MockFunctionReplaceVisitor(Opcodes.ASM7, cw, method.getName(),
-                Type.getMethodDescriptor(method), mockMethod);
-        cr.accept(cv, 0);
-        return cw.toByteArray();
-    }
-
-    //Get all mocked functions in a class
-    private static void populateClassNameVsFunctionToMockMap(Map<String, List<String>> classVsMockFunctionsMap,
-                                                             Map<String, String> mockFunctionMap) {
-        for (Map.Entry<String, String> entry : mockFunctionMap.entrySet()) {
-            String key = entry.getKey();
-            String functionToMockClassName;
-            String functionToMock;
-            if (key.indexOf(MOCK_LEGACY_DELIMITER) == -1) {
-                functionToMockClassName = key.substring(0, key.indexOf(MOCK_FN_DELIMITER));
-                functionToMock = key.substring(key.indexOf(MOCK_FN_DELIMITER));
-            } else if (key.indexOf(MOCK_FN_DELIMITER) == -1) {
-                functionToMockClassName = key.substring(0, key.indexOf(MOCK_LEGACY_DELIMITER));
-                functionToMock = key.substring(key.indexOf(MOCK_LEGACY_DELIMITER));
-            } else {
-                if (key.indexOf(MOCK_FN_DELIMITER) < key.indexOf(MOCK_LEGACY_DELIMITER)) {
-                    functionToMockClassName = key.substring(0, key.indexOf(MOCK_FN_DELIMITER));
-                    functionToMock = key.substring(key.indexOf(MOCK_FN_DELIMITER));
-                } else {
-                    functionToMockClassName = key.substring(0, key.indexOf(MOCK_LEGACY_DELIMITER));
-                    functionToMock = key.substring(key.indexOf(MOCK_LEGACY_DELIMITER));
-                }
-            }
-            functionToMock = functionToMock.replaceAll("\\\\", "");
-            classVsMockFunctionsMap.computeIfAbsent(functionToMockClassName,
-                    k -> new ArrayList<>()).add(functionToMock);
-        }
     }
 
     @Override
@@ -373,7 +223,7 @@ public class RunNativeImageTestTask implements Task {
         for (Map<String, TestSuite> testSuiteMap : testSuiteMapEntries) {
             try {
                 Path nativeConfigPath = target.getNativeConfigPath();
-                createReflectConfig(nativeConfigPath, project.currentPackage(), testSuiteMap);
+                NativeUtils.createReflectConfig(nativeConfigPath, project.currentPackage(), testSuiteMap);
             } catch (IOException e) {
                 throw createLauncherException("error while generating the necessary graalvm reflection config ", e);
             }
@@ -383,7 +233,7 @@ public class RunNativeImageTestTask implements Task {
                 TestSuite testSuite = testSuiteMap.values().toArray(new TestSuite[0])[0];
                 String moduleName = testSuite.getPackageID();
                 try {
-                    modifyJarForFunctionMock(testSuite, target, moduleName);
+                    NativeUtils.modifyJarForFunctionMock(testSuite, target, moduleName);
                 } catch (IOException e) {
                     throw createLauncherException("error occurred while running tests", e);
                 }
@@ -464,7 +314,7 @@ public class RunNativeImageTestTask implements Task {
                                             Map<String, TestSuite> testSuiteMap)
             throws IOException, InterruptedException {
         String packageName = currentPackage.packageName().toString();
-        String classPath = getClassPath(testSuiteMap);
+        String classPath = NativeUtils.getClassPath(testSuiteMap);
 
         String jacocoAgentJarPath = "";
         String nativeImageCommand = System.getenv("GRAALVM_HOME");
@@ -476,7 +326,7 @@ public class RunNativeImageTestTask implements Task {
                         "https://ballerina.io/learn/build-the-executable-locally/#configure-graalvm");
             }
             nativeImageCommand += File.separator + BIN_DIR_NAME + File.separator
-                    + (OS.contains("win") ? "native-image.cmd" : "native-image");
+                    + (NativeUtils.OS.contains("win") ? "native-image.cmd" : "native-image");
 
             File commandExecutable = Paths.get(nativeImageCommand).toFile();
             if (!commandExecutable.exists()) {
@@ -514,11 +364,11 @@ public class RunNativeImageTestTask implements Task {
 
 
         // set name and path
-        nativeArgs.add("-H:Name=" + addQuotationMarkToString(packageName));
-        nativeArgs.add("-H:Path=" + convertWinPathToUnixFormat(addQuotationMarkToString(nativeTargetPath.toString())));
+        nativeArgs.add("-H:Name=" + NativeUtils.addQuotationMarkToString(packageName));
+        nativeArgs.add("-H:Path=" + NativeUtils.convertWinPathToUnixFormat(NativeUtils.addQuotationMarkToString(nativeTargetPath.toString())));
 
         // native-image configs
-        nativeArgs.add("-H:ReflectionConfigurationFiles=" + convertWinPathToUnixFormat(addQuotationMarkToString(
+        nativeArgs.add("-H:ReflectionConfigurationFiles=" + NativeUtils.convertWinPathToUnixFormat(NativeUtils.addQuotationMarkToString(
                 nativeConfigPath.resolve("reflect-config.json").toString())));
         nativeArgs.add("--no-fallback");
 
@@ -549,7 +399,10 @@ public class RunNativeImageTestTask implements Task {
             cmdArgs.add(generatedImagePath);
 
             // Test Runner Class arguments
-            cmdArgs.add(target.path().toString());                                  // 0
+            cmdArgs.add(Boolean.toString(false));                                 //0
+            cmdArgs.add(target.path().resolve(CACHE_DIR).resolve(TESTS_CACHE_DIR_NAME)
+                    .resolve(TESTERINA_TEST_SUITE).toString());
+            cmdArgs.add(target.path().toString());
             cmdArgs.add(jacocoAgentJarPath);
             cmdArgs.add(Boolean.toString(report));
             cmdArgs.add(Boolean.toString(coverage));
@@ -557,7 +410,7 @@ public class RunNativeImageTestTask implements Task {
             cmdArgs.add(this.disableGroupList != null ? this.disableGroupList : "");
             cmdArgs.add(this.singleExecTests != null ? this.singleExecTests : "");
             cmdArgs.add(Boolean.toString(isRerunTestExecution));
-            cmdArgs.add(Boolean.toString(listGroups));                              // 8
+            cmdArgs.add(Boolean.toString(listGroups));                              // 10             // 8
             cmdArgs.add(Boolean.toString(isParallelExecution));
 
             builder.command(cmdArgs.toArray(new String[0]));
@@ -572,18 +425,6 @@ public class RunNativeImageTestTask implements Task {
             return 1;
         }
     }
-
-    private String addQuotationMarkToString(String word) {
-        return "\"" + word + "\"";
-    }
-
-    private String convertWinPathToUnixFormat(String path) {
-        if (OS.contains("win")) {
-            path = path.replace("\\", "/");
-        }
-        return path;
-    }
-
 
     private void validateResourcesWithinJar(Map<String, TestSuite> testSuiteMap, String packageName)
             throws IOException {
@@ -612,152 +453,8 @@ public class RunNativeImageTestTask implements Task {
         }
     }
 
-
-    private String getClassPath(Map<String, TestSuite> testSuiteMap) {
-        List<String> dependencies = new ArrayList<>();
-        for (Map.Entry<String, TestSuite> testSuiteEntry : testSuiteMap.entrySet()) {
-            dependencies.addAll(testSuiteEntry.getValue().getTestExecutionDependencies());
-
-        }
-        dependencies = dependencies.stream().distinct().collect(Collectors.toList());
-        dependencies = dependencies.stream().map((x) -> convertWinPathToUnixFormat(addQuotationMarkToString(x)))
-                .collect(Collectors.toList());
-
-        StringJoiner classPath = new StringJoiner(File.pathSeparator);
-        dependencies.stream().forEach(classPath::add);
-        return classPath.toString();
-    }
-
-    private void modifyJarForFunctionMock(TestSuite testSuite, Target target, String moduleName) throws IOException {
-        String testJarName = testSuite.getOrgName() + HYPHEN + moduleName + HYPHEN +
-                testSuite.getVersion() + HYPHEN + TESTABLE + JAR_EXTENSION;
-        String testJarPath = "";
-        String modifiedJarName = "";
-        String mainJarPath = "";
-        String mainJarName = "";
-
-        if (testSuite.getMockFunctionNamesMap().isEmpty()) {
-            return;
-        }
-
-        //Add testable jar path to classloader URLs
-        List<String> testExecutionDependencies = testSuite.getTestExecutionDependencies();
-        List<String> classLoaderUrlList = new ArrayList<>();
-        for (String testExecutionDependency : testExecutionDependencies) {
-            if (testExecutionDependency.endsWith(testJarName)) {
-                testJarPath = testExecutionDependency;
-                classLoaderUrlList.add(testJarPath);
-            }
-        }
-
-        ClassLoader classLoader = null;
-
-        //Extract the className vs mocking functions list
-        Map<String, List<String>> classVsMockFunctionsMap = new HashMap<>();
-        Map<String, String> mockFunctionMap = testSuite.getMockFunctionNamesMap();
-        populateClassNameVsFunctionToMockMap(classVsMockFunctionsMap, mockFunctionMap);
-
-        //Extract a mapping between classes and corresponding module jar
-        Map<String, List<String>> mainJarVsClassMapping = new HashMap<>();
-        for (Map.Entry<String, List<String>> classVsMockFunctionsEntry : classVsMockFunctionsMap.entrySet()) {
-            String className = classVsMockFunctionsEntry.getKey();
-            String[] classMetaData = className.split("\\.");
-            mainJarName = classMetaData[0] + HYPHEN + classMetaData[1].replace(DOT_REPLACER, DOT) +
-                    HYPHEN + classMetaData[2];
-
-            if (mainJarVsClassMapping.containsKey(mainJarName)) {
-                mainJarVsClassMapping.get(mainJarName).add(className);
-            } else {
-                List<String> classList = new ArrayList<>();
-                classList.add(className);
-                mainJarVsClassMapping.put(mainJarName, classList);
-            }
-        }
-
-        //Modify classes within module jar based on above mapping
-        for (Map.Entry<String, List<String>> mainJarVsClassEntry : mainJarVsClassMapping.entrySet()) {
-
-            mainJarName = mainJarVsClassEntry.getKey();
-            modifiedJarName = mainJarName + HYPHEN + MODIFIED + JAR_EXTENSION;
-
-            for (String testExecutionDependency : testExecutionDependencies) {
-                if (testExecutionDependency.contains(mainJarName) && !testExecutionDependency.contains(TESTABLE)) {
-                    mainJarPath = testExecutionDependency;
-                }
-            }
-            //Add module jar path to classloader URLs
-            classLoaderUrlList.add(mainJarPath);
-            classLoader = AccessController.doPrivileged(
-                    (PrivilegedAction<URLClassLoader>) () -> new URLClassLoader(getURLList(classLoaderUrlList).
-                            toArray(new URL[0]), ClassLoader.getSystemClassLoader()));
-
-            //Modify classes within jar
-            Map<String, byte[]> modifiedClassDef = new HashMap<>();
-            for (String className : mainJarVsClassEntry.getValue()) {
-                List<String> functionNamesList = classVsMockFunctionsMap.get(className);
-                byte[] classFile = getModifiedClassBytes(className, functionNamesList, testSuite, classLoader);
-                modifiedClassDef.put(className, classFile);
-            }
-
-            //Load all classes within module jar
-            Map<String, byte[]> unmodifiedFiles = loadUnmodifiedFilesWithinJar(mainJarPath);
-            String modifiedJarPath = (target.path().resolve(CACHE_DIR).resolve(testSuite.getOrgName()).resolve
-                    (testSuite.getPackageName()).resolve(testSuite.getVersion()).resolve(JAVA_17_DIR)).toString()
-                    + PATH_SEPARATOR + modifiedJarName;
-            //Dump modified jar
-            dumpJar(modifiedClassDef, unmodifiedFiles, modifiedJarPath);
-
-            testExecutionDependencies.remove(mainJarPath);
-            testExecutionDependencies.add(modifiedJarPath);
-        }
-    }
-
-    //Replace unmodified classes with corresponding modified classes and dump jar
-    private void dumpJar(Map<String, byte[]> modifiedClassDefs, Map<String, byte[]> unmodifiedFiles,
-                         String modifiedJarPath) throws IOException {
-        List<String> duplicatePaths = new ArrayList<>();
-        try (JarOutputStream jarOutputStream = new JarOutputStream(new FileOutputStream(modifiedJarPath))) {
-            for (Map.Entry<String, byte[]> modifiedClassDef : modifiedClassDefs.entrySet()) {
-                if (modifiedClassDef.getValue().length > 0) {
-                    String entry = modifiedClassDef.getKey();
-                    String path = entry.replaceAll("\\.", PATH_SEPARATOR) + CLASS_EXTENSION;
-                    duplicatePaths.add(path);
-                    jarOutputStream.putNextEntry(new ZipEntry(path));
-                    jarOutputStream.write(modifiedClassDefs.get(entry));
-                    jarOutputStream.closeEntry();
-                }
-            }
-            for (Map.Entry<String, byte[]> unmodifiedFile : unmodifiedFiles.entrySet()) {
-                String entry = unmodifiedFile.getKey();
-                if (!duplicatePaths.contains(entry)) {
-                    jarOutputStream.putNextEntry(new ZipEntry(entry));
-                    jarOutputStream.write(unmodifiedFiles.get(entry));
-                    jarOutputStream.closeEntry();
-                }
-            }
-        }
-    }
-
-    private Map<String, byte[]> loadUnmodifiedFilesWithinJar(String mainJarPath)
-            throws IOException {
-        Map<String, byte[]> unmodifiedFiles = new HashMap<String, byte[]>();
-        File jarFile = new File(mainJarPath);
-        ZipInputStream jarInputStream = new ZipInputStream(new FileInputStream(jarFile));
-        ZipEntry entry;
-        while ((entry = jarInputStream.getNextEntry()) != null) {
-            String path = entry.getName();
-            if (!entry.isDirectory()) {
-                byte[] bytes = IOUtils.toByteArray(jarInputStream);
-                unmodifiedFiles.put(path, bytes);
-            }
-            jarInputStream.closeEntry();
-        }
-        jarInputStream.close();
-        return unmodifiedFiles;
-    }
-
     private String getGeneratedImageExtension() {
-        if (OS.contains("win")) {
+        if (NativeUtils.OS.contains("win")) {
             return DOT + WIN_EXEC_EXT;
         }
         return "";
