+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/utils/NativeUtils.java
@@ -22,14 +22,15 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
 import io.ballerina.projects.Package;
+import io.ballerina.projects.internal.model.Target;
 import io.ballerina.runtime.internal.util.RuntimeUtils;
+import org.apache.commons.compress.utils.IOUtils;
+import org.ballerinalang.test.runtime.entity.MockFunctionReplaceVisitor;
 import org.ballerinalang.test.runtime.entity.TestSuite;
+import org.ballerinalang.test.runtime.util.TesterinaUtils;
+import org.objectweb.asm.*;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.Writer;
+import java.io.*;
 import java.lang.reflect.Method;
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -41,18 +42,18 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.security.AccessController;
 import java.security.PrivilegedAction;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.jar.JarOutputStream;
+import java.util.stream.Collectors;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 
+import static io.ballerina.cli.launcher.LauncherUtils.createLauncherException;
 import static io.ballerina.identifier.Utils.encodeNonFunctionIdentifier;
 import static io.ballerina.runtime.api.constants.RuntimeConstants.FILE_NAME_PERIOD_SEPARATOR;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.ANON_ORG;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.DOT;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.MOCK_FN_DELIMITER;
-import static org.ballerinalang.test.runtime.util.TesterinaConstants.MOCK_LEGACY_DELIMITER;
+import static java.util.Objects.requireNonNull;
+import static org.ballerinalang.test.runtime.util.TesterinaConstants.*;
+import static org.ballerinalang.test.runtime.util.TesterinaConstants.PATH_SEPARATOR;
 
 /**
  * Utility functions and classes for test native-image generation.
@@ -62,6 +63,7 @@ import static org.ballerinalang.test.runtime.util.TesterinaConstants.MOCK_LEGACY
 public class NativeUtils {
     private static final String MODULE_INIT_CLASS_NAME = "$_init";
     private static final String TEST_EXEC_FUNCTION = "__execute__";
+    public static final String OS = System.getProperty("os.name").toLowerCase(Locale.getDefault());
 
     private static final ReflectConfigClassMethod REFLECTION_CONFIG_EXECUTE_METHOD = new ReflectConfigClassMethod(
             TEST_EXEC_FUNCTION, new String[]{"io.ballerina.runtime.internal.scheduling.Strand",
@@ -317,6 +319,279 @@ public class NativeUtils {
             className = encodeNonFunctionIdentifier(orgName) + "." + className;
         }
         return className;
+    }
+
+    public static void modifyJarForFunctionMock(TestSuite testSuite, Target target, String moduleName)
+            throws IOException {
+        String testJarName = testSuite.getOrgName() + HYPHEN + moduleName + HYPHEN +
+                testSuite.getVersion() + HYPHEN + TESTABLE + JAR_EXTENSION;
+        String testJarPath = "";
+        String modifiedJarName = "";
+        String mainJarPath = "";
+        String mainJarName = "";
+
+        if (testSuite.getMockFunctionNamesMap().isEmpty()) {
+            return;
+        }
+
+        //Add testable jar path to classloader URLs
+        List<String> testExecutionDependencies = testSuite.getTestExecutionDependencies();
+        List<String> classLoaderUrlList = new ArrayList<>();
+        for (String testExecutionDependency : testExecutionDependencies) {
+            if (testExecutionDependency.endsWith(testJarName)) {
+                testJarPath = testExecutionDependency;
+                classLoaderUrlList.add(testJarPath);
+            }
+        }
+
+        ClassLoader classLoader = null;
+
+        //Extract the className vs mocking functions list
+        Map<String, List<String>> classVsMockFunctionsMap = new HashMap<>();
+        Map<String, String> mockFunctionMap = testSuite.getMockFunctionNamesMap();
+        populateClassNameVsFunctionToMockMap(classVsMockFunctionsMap, mockFunctionMap);
+
+        //Extract a mapping between classes and corresponding module jar
+        Map<String, List<String>> mainJarVsClassMapping = new HashMap<>();
+        for (Map.Entry<String, List<String>> classVsMockFunctionsEntry : classVsMockFunctionsMap.entrySet()) {
+            String className = classVsMockFunctionsEntry.getKey();
+            String[] classMetaData = className.split("\\.");
+            mainJarName = classMetaData[0] + HYPHEN + classMetaData[1].replace(DOT_REPLACER, DOT) +
+                    HYPHEN + classMetaData[2];
+
+            if (mainJarVsClassMapping.containsKey(mainJarName)) {
+                mainJarVsClassMapping.get(mainJarName).add(className);
+            } else {
+                List<String> classList = new ArrayList<>();
+                classList.add(className);
+                mainJarVsClassMapping.put(mainJarName, classList);
+            }
+        }
+
+        //Modify classes within module jar based on above mapping
+        for (Map.Entry<String, List<String>> mainJarVsClassEntry : mainJarVsClassMapping.entrySet()) {
+
+            mainJarName = mainJarVsClassEntry.getKey();
+            modifiedJarName = mainJarName + HYPHEN + MODIFIED + JAR_EXTENSION;
+
+            for (String testExecutionDependency : testExecutionDependencies) {
+                if (testExecutionDependency.contains(mainJarName) && !testExecutionDependency.contains(TESTABLE)) {
+                    mainJarPath = testExecutionDependency;
+                    break;
+                }
+            }
+            //Add module jar path to classloader URLs
+            classLoaderUrlList.add(mainJarPath);
+            classLoader = AccessController.doPrivileged(
+                    (PrivilegedAction<URLClassLoader>) () -> new URLClassLoader(getURLList(classLoaderUrlList).
+                            toArray(new URL[0]), ClassLoader.getSystemClassLoader()));
+
+            //Modify classes within jar
+            Map<String, byte[]> modifiedClassDef = new HashMap<>();
+            for (String className : mainJarVsClassEntry.getValue()) {
+                List<String> functionNamesList = classVsMockFunctionsMap.get(className);
+                byte[] classFile = getModifiedClassBytes(className, functionNamesList, testSuite, classLoader);
+                modifiedClassDef.put(className, classFile);
+            }
+
+            //Load all classes within module jar
+            Map<String, byte[]> unmodifiedFiles = loadUnmodifiedFilesWithinJar(mainJarPath);
+            String modifiedJarPath = (target.path().resolve(CACHE_DIR).resolve(testSuite.getOrgName()).resolve
+                    (testSuite.getPackageName()).resolve(testSuite.getVersion()).resolve(JAVA_17_DIR)).toString()
+                    + PATH_SEPARATOR + modifiedJarName;
+            //Dump modified jar
+            dumpJar(modifiedClassDef, unmodifiedFiles, modifiedJarPath);
+
+            testExecutionDependencies.remove(mainJarPath);
+            testExecutionDependencies.add(modifiedJarPath);
+        }
+    }
+
+    //Replace unmodified classes with corresponding modified classes and dump jar
+    private static void dumpJar(Map<String, byte[]> modifiedClassDefs, Map<String, byte[]> unmodifiedFiles,
+                         String modifiedJarPath) throws IOException {
+        List<String> duplicatePaths = new ArrayList<>();
+        try (JarOutputStream jarOutputStream = new JarOutputStream(new FileOutputStream(modifiedJarPath))) {
+            for (Map.Entry<String, byte[]> modifiedClassDef : modifiedClassDefs.entrySet()) {
+                if (modifiedClassDef.getValue().length > 0) {
+                    String entry = modifiedClassDef.getKey();
+                    String path = entry.replaceAll("\\.", PATH_SEPARATOR) + CLASS_EXTENSION;
+                    duplicatePaths.add(path);
+                    jarOutputStream.putNextEntry(new ZipEntry(path));
+                    jarOutputStream.write(modifiedClassDefs.get(entry));
+                    jarOutputStream.closeEntry();
+                }
+            }
+            for (Map.Entry<String, byte[]> unmodifiedFile : unmodifiedFiles.entrySet()) {
+                String entry = unmodifiedFile.getKey();
+                if (!duplicatePaths.contains(entry)) {
+                    jarOutputStream.putNextEntry(new ZipEntry(entry));
+                    jarOutputStream.write(unmodifiedFiles.get(entry));
+                    jarOutputStream.closeEntry();
+                }
+            }
+        }
+    }
+
+    private static Map<String, byte[]> loadUnmodifiedFilesWithinJar(String mainJarPath)
+            throws IOException {
+        Map<String, byte[]> unmodifiedFiles = new HashMap<String, byte[]>();
+        File jarFile = new File(mainJarPath);
+        ZipInputStream jarInputStream = new ZipInputStream(new FileInputStream(jarFile));
+        ZipEntry entry;
+        while ((entry = jarInputStream.getNextEntry()) != null) {
+            String path = entry.getName();
+            if (!entry.isDirectory()) {
+                byte[] bytes = IOUtils.toByteArray(jarInputStream);
+                unmodifiedFiles.put(path, bytes);
+            }
+            jarInputStream.closeEntry();
+        }
+        jarInputStream.close();
+        return unmodifiedFiles;
+    }
+
+    //Get all mocked functions in a class
+    private static void populateClassNameVsFunctionToMockMap(Map<String, List<String>> classVsMockFunctionsMap,
+                                                             Map<String, String> mockFunctionMap) {
+        for (Map.Entry<String, String> entry : mockFunctionMap.entrySet()) {
+            String key = entry.getKey();
+            String functionToMockClassName;
+            String functionToMock;
+            if (key.indexOf(MOCK_LEGACY_DELIMITER) == -1) {
+                functionToMockClassName = key.substring(0, key.indexOf(MOCK_FN_DELIMITER));
+                functionToMock = key.substring(key.indexOf(MOCK_FN_DELIMITER));
+            } else if (key.indexOf(MOCK_FN_DELIMITER) == -1) {
+                functionToMockClassName = key.substring(0, key.indexOf(MOCK_LEGACY_DELIMITER));
+                functionToMock = key.substring(key.indexOf(MOCK_LEGACY_DELIMITER));
+            } else {
+                if (key.indexOf(MOCK_FN_DELIMITER) < key.indexOf(MOCK_LEGACY_DELIMITER)) {
+                    functionToMockClassName = key.substring(0, key.indexOf(MOCK_FN_DELIMITER));
+                    functionToMock = key.substring(key.indexOf(MOCK_FN_DELIMITER));
+                } else {
+                    functionToMockClassName = key.substring(0, key.indexOf(MOCK_LEGACY_DELIMITER));
+                    functionToMock = key.substring(key.indexOf(MOCK_LEGACY_DELIMITER));
+                }
+            }
+            functionToMock = functionToMock.replaceAll("\\\\", "");
+            classVsMockFunctionsMap.computeIfAbsent(functionToMockClassName,
+                    k -> new ArrayList<>()).add(functionToMock);
+        }
+    }
+
+    private static byte[] getModifiedClassBytes(String className, List<String> functionNames, TestSuite suite,
+                                                ClassLoader classLoader) {
+        Class<?> functionToMockClass;
+        try {
+            functionToMockClass = classLoader.loadClass(className);
+        } catch (Throwable e) {
+            throw createLauncherException("failed to load class: " + className);
+        }
+
+        byte[] classFile = new byte[0];
+        boolean readFromBytes = false;
+        for (Method method1 : functionToMockClass.getDeclaredMethods()) {
+            if (functionNames.contains(MOCK_FN_DELIMITER + method1.getName())) {
+                String desugaredMockFunctionName = MOCK_FUNC_NAME_PREFIX + method1.getName();
+                String testClassName = TesterinaUtils.getQualifiedClassName(suite.getOrgName(),
+                        suite.getTestPackageID(), suite.getVersion(),
+                        suite.getPackageID().replace(DOT, FILE_NAME_PERIOD_SEPARATOR));
+                Class<?> testClass;
+                try {
+                    testClass = classLoader.loadClass(testClassName);
+                } catch (Throwable e) {
+                    throw createLauncherException("failed to prepare " + testClassName + " for mocking reason:" +
+                            e.getMessage());
+                }
+                for (Method method2 : testClass.getDeclaredMethods()) {
+                    if (method2.getName().equals(desugaredMockFunctionName)) {
+                        if (!readFromBytes) {
+                            classFile = replaceMethodBody(method1, method2);
+                            readFromBytes = true;
+                        } else {
+                            classFile = replaceMethodBody(classFile, method1, method2);
+                        }
+                    }
+                }
+            } else if (functionNames.contains(MOCK_LEGACY_DELIMITER + method1.getName())) {
+                String key = className + MOCK_LEGACY_DELIMITER + method1.getName();
+                String mockFunctionName = suite.getMockFunctionNamesMap().get(key);
+                if (mockFunctionName != null) {
+                    String mockFunctionClassName = suite.getTestUtilityFunctions().get(mockFunctionName);
+                    Class<?> mockFunctionClass;
+                    try {
+                        mockFunctionClass = classLoader.loadClass(mockFunctionClassName);
+                    } catch (ClassNotFoundException e) {
+                        throw createLauncherException("failed to prepare " + mockFunctionClassName +
+                                " for mocking reason:" + e.getMessage());
+                    }
+                    for (Method method2 : mockFunctionClass.getDeclaredMethods()) {
+                        if (method2.getName().equals(mockFunctionName)) {
+                            if (!readFromBytes) {
+                                classFile = replaceMethodBody(method1, method2);
+                                readFromBytes = true;
+                            } else {
+                                classFile = replaceMethodBody(classFile, method1, method2);
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        return classFile;
+    }
+
+    private static byte[] replaceMethodBody(Method method, Method mockMethod) {
+        Class<?> clazz = method.getDeclaringClass();
+        ClassReader cr;
+        try {
+            InputStream ins;
+            ins = clazz.getResourceAsStream(clazz.getSimpleName() + CLASS_EXTENSION);
+            cr = new ClassReader(requireNonNull(ins));
+        } catch (IOException e) {
+            throw createLauncherException("failed to get the class reader object for the class "
+                    + clazz.getSimpleName());
+        }
+        ClassWriter cw = new ClassWriter(cr, ClassWriter.COMPUTE_FRAMES | ClassWriter.COMPUTE_MAXS);
+        ClassVisitor cv = new MockFunctionReplaceVisitor(Opcodes.ASM7, cw, method.getName(),
+                Type.getMethodDescriptor(method), mockMethod);
+        cr.accept(cv, 0);
+        return cw.toByteArray();
+    }
+
+    private static byte[] replaceMethodBody(byte[] classFile, Method method, Method mockMethod) {
+        ClassReader cr = new ClassReader(classFile);
+        ClassWriter cw = new ClassWriter(cr, ClassWriter.COMPUTE_FRAMES | ClassWriter.COMPUTE_MAXS);
+        ClassVisitor cv = new MockFunctionReplaceVisitor(Opcodes.ASM7, cw, method.getName(),
+                Type.getMethodDescriptor(method), mockMethod);
+        cr.accept(cv, 0);
+        return cw.toByteArray();
+    }
+
+    public static String getClassPath(Map<String, TestSuite> testSuiteMap) {
+        List<String> dependencies = new ArrayList<>();
+        for (Map.Entry<String, TestSuite> testSuiteEntry : testSuiteMap.entrySet()) {
+            dependencies.addAll(testSuiteEntry.getValue().getTestExecutionDependencies());
+
+        }
+        dependencies = dependencies.stream().distinct().collect(Collectors.toList());
+        dependencies = dependencies.stream().map((x) -> convertWinPathToUnixFormat(addQuotationMarkToString(x)))
+                .collect(Collectors.toList());
+
+        StringJoiner classPath = new StringJoiner(File.pathSeparator);
+        dependencies.stream().forEach(classPath::add);
+        return classPath.toString();
+    }
+
+    public static String addQuotationMarkToString(String word) {
+        return "\"" + word + "\"";
+    }
+
+    public static String convertWinPathToUnixFormat(String path) {
+        if (OS.contains("win")) {
+            path = path.replace("\\", "/");
+        }
+        return path;
     }
 
     private static class ReflectConfigClass {
